Move helper methods to util
