--- app/helpers/index.js
@@ -1,5 +1,0 @@
-'use strict'
-module.exports = {
-  dataHelper: require('./data'),
-  timeHelper: require('./time')
-}
