+++ app/util/index.js
@@ -2,5 +2,6 @@
 module.exports = {
   argumentUtil: require('./argument'),
   Constants: require('./constants'),
+  timeUtil: require('./time'),
   util: require('./util')
 }
